Refactor authentication flow: Remove Microsoft and Google login components, consolidate Google authentication logic, and update environment configuration. Implement CourseSearch page with API integration and user session management. Enhance styling for login and course search interfaces.
