--- frontend-web/.env.development
@@ -1,3 +1,0 @@
-VITE_API_BASE=http://localhost:8080
-VITE_AZURE_CLIENT_ID=REPLACE_WITH_YOUR_AZURE_CLIENT_ID
-VITE_GOOGLE_CLIENT_ID=REPLACE_WITH_YOUR_GOOGLE_CLIENT_ID
