+++ frontend-web/.gitignore
@@ -13,8 +13,8 @@ dist-ssr
 *.local
 
 # Environment variables
-.env*
-!.env.example
+env/.env.development
+env/.env.production
 
 # Editor directories and files
 .vscode/*
