+++ frontend-web/src/config/index.js
@@ -1,0 +1,1 @@
+export * from './googleAuthConfig.js'
