+++ frontend-web/src/main.jsx
@@ -1,18 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { PublicClientApplication } from '@azure/msal-browser'
-import { MsalProvider } from '@azure/msal-react'
-import { msalConfig } from './authConfig.js'
 import './index.css'
 import App from './App.jsx'
 
-// Create MSAL instance
-const msalInstance = new PublicClientApplication(msalConfig)
-
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <MsalProvider instance={msalInstance}>
-      <App />
-    </MsalProvider>
+    <App />
   </StrictMode>,
 )
