+++ frontend-web/src/pages/CourseSearch/index.js
@@ -1,0 +1,1 @@
+export { default } from './CourseSearch.jsx'
