+++ frontend-web/src/pages/Login/Login.jsx
@@ -1,15 +1,8 @@
 import { useState, useEffect } from 'react'
-import { useMsal } from '@azure/msal-react'
-import { loginRequest } from './authConfig.js'
-import { initializeGoogleAuth, signInWithGoogle } from './googleAuthConfig.js'
+import { initializeGoogleAuth, signInWithGoogle } from '../../config/googleAuthConfig.js'
 import './Login.css'
 
 function Login({ onLogin }) {
-  const { instance } = useMsal()
-  const [credentials, setCredentials] = useState({
-    email: '',
-    password: ''
-  })
   const [isGoogleReady, setIsGoogleReady] = useState(false)
   const [errorMessage, setErrorMessage] = useState('')
   const [showGsuEmailModal, setShowGsuEmailModal] = useState(false)
@@ -30,56 +23,6 @@ function Login({ onLogin }) {
     }
     setupGoogle()
   }, [])
-
-  const handleChange = (e) => {
-    setCredentials({
-      ...credentials,
-      [e.target.name]: e.target.value
-    })
-  }
-
-  const handleSubmit = (e) => {
-    e.preventDefault()
-    setErrorMessage('')
-    console.log('Login attempt:', credentials)
-    // For now, just redirect to main form
-    onLogin()
-  }
-
-  const handleMicrosoftLogin = async () => {
-    try {
-      setErrorMessage('')
-      console.log('Starting Microsoft login...')
-      const response = await instance.loginPopup(loginRequest)
-      console.log('Microsoft login successful:', response)
-      
-      // Extract user info
-      const { account } = response
-      const userEmail = account.username
-      
-      // Validate GSU email domain
-      if (!userEmail.endsWith('@student.gsu.edu') && !userEmail.endsWith('@gsu.edu')) {
-        console.error('Invalid email domain. Please use your GSU email address.')
-        setErrorMessage('Please sign in with your GSU email address (@student.gsu.edu or @gsu.edu)')
-        
-        // Log out the user since they don't have the right email domain
-        await instance.logoutPopup()
-        return
-      }
-      
-      console.log('Valid GSU user:', {
-        name: account.name,
-        email: userEmail,
-        id: account.homeAccountId
-      })
-      
-      // Proceed to main app
-      onLogin()
-    } catch (error) {
-      console.error('Microsoft login failed:', error)
-      setErrorMessage('Microsoft login failed. Please try again.')
-    }
-  }
 
   const handleGoogleLogin = async () => {
     try {
@@ -153,10 +96,11 @@ function Login({ onLogin }) {
       <div className="login-card">
         <header className="login-header">
           <h1>PantherWatch</h1>
-          <p>Sign in to search GSU courses</p>
+          <p>Georgia State University Course Search</p>
+          <p className="login-subtitle">Sign in with your Google account to get started</p>
         </header>
         
-        {/* OAuth Login Buttons */}
+        {/* Google Login Button */}
         <div className="oauth-section">
           <button type="button" className="oauth-btn google-btn" onClick={handleGoogleLogin}>
             <svg className="oauth-icon" viewBox="0 0 24 24" fill="currentColor">
@@ -165,14 +109,7 @@ function Login({ onLogin }) {
               <path d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z" fill="#FBBC05"/>
               <path d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z" fill="#EA4335"/>
             </svg>
-            Sign in with Google (Personal Email)
-          </button>
-          
-          <button type="button" className="oauth-btn microsoft-btn" onClick={handleMicrosoftLogin}>
-            <svg className="oauth-icon" viewBox="0 0 24 24" fill="currentColor">
-              <path d="M11.4 24H0V12.6h11.4V24zM24 24H12.6V12.6H24V24zM11.4 11.4H0V0h11.4v11.4zM24 11.4H12.6V0H24v11.4z"/>
-            </svg>
-            Sign in with Microsoft (Student Email) - WIP
+            Continue with Google
           </button>
         </div>
         
@@ -183,45 +120,9 @@ function Login({ onLogin }) {
           </div>
         )}
         
-        {/* Divider */}
-        <div className="login-divider">
-          <span>or</span>
-        </div>
-        
-        <form className="login-form" onSubmit={handleSubmit}>
-          <div className="field-group">
-            <label htmlFor="email">Email</label>
-            <input
-              id="email"
-              name="email"
-              type="email"
-              value={credentials.email}
-              onChange={handleChange}
-              placeholder="your.email@student.gsu.edu"
-              required
-            />
-          </div>
-          
-          <div className="field-group">
-            <label htmlFor="password">Password</label>
-            <input
-              id="password"
-              name="password"
-              type="password"
-              value={credentials.password}
-              onChange={handleChange}
-              placeholder="Enter your password"
-              required
-            />
-          </div>
-          
-          <button type="submit" className="login-btn">
-            Login
-          </button>
-        </form>
-        
         <footer className="login-footer">
-          <p>Access GSU course information and availability</p>
+          <p>Access real-time GSU course information and availability</p>
+          <p className="login-note">You'll be asked for your GSU email after signing in</p>
         </footer>
       </div>
